requirements.txt
fastapi
uvicorn
aiofiles
psycopg2-binary
aioping
icmplib
SQLAlchemy
asyncpg
pydantic_settings